Separate the program arguments handling and console input/output from the theme conversion code. Add a new test project for the converter.
