+++ ThemeConverter/ThemeConverterTests/TestFiles/Invalid_NoColors.json
@@ -1,0 +1,6 @@
+{
+	"$schema": "vscode://schemas/color-theme",
+	"type": "light",
+	"colors": {
+	}
+}
